Inline h.client.embed_context into h.views.client.embed
The `h.client` module was created for rendering the app.html and
embed.js files during the browser extension build. In that environment
there is no request object.

Since the browser extension build has now been moved to a separate repo,
there is no need for having `h.client.embed_context` as a separate
function from the view itself.
